fix!: 사람이 더 잘 읽을 수 있는 인터페이스로 변경 (#6)
